AZ-23372: moved format_coordinates to utils to reduce duplicate code
